Move putCursorAtEnd to its own file to avoid mini-racer issues
