+++ app/assets/javascripts/discourse/lib/utilities.js
@@ -1,6 +1,5 @@
 import { escape } from "pretty-text/sanitizer";
 import toMarkdown from "discourse/lib/to-markdown";
-import positioningWorkaround from "discourse/lib/safari-hacks";
 
 const homepageSelector = "meta[name=discourse_current_homepage]";
 
@@ -443,20 +442,6 @@ export function inCodeBlock(text, pos) {
   }
 
   return result;
-}
-
-export function putCursorAtEnd(element) {
-  if (isAppleDevice() && positioningWorkaround.touchstartEvent) {
-    positioningWorkaround.touchstartEvent(element);
-  } else {
-    element.focus();
-  }
-
-  const len = element.value.length;
-  element.setSelectionRange(len, len);
-
-  // Scroll to the bottom, in case we're in a tall textarea
-  element.scrollTop = 999999;
 }
 
 // This prevents a mini racer crash
